+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Convert "Go to previous post" button into link.
 - Renamed "Add feed" button to "Create feed".
 - Renamed "View replies" link to "View conversation".
+- Moved "Delete feed" button to feed configuration page.
 
 ### Fixed